refactor: isolate Playground runtime orchestration
